pyproject.toml
[tool.poetry]
name = "breaking-the-bad-habit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
generics = "^7.0.0"


[tool.poetry.group.dependence.dependencies]
django = "^5.1.4"
pillow = "^11.0.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
ipython = "^8.30.0"
black = "^24.10.0"


[tool.poetry.group.dependence_drf.dependencies]
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dependence_test.dependencies]
coverage = "^7.6.9"


[tool.poetry.group.dependence_task.dependencies]
redis = "^5.2.1"
celery = "^5.4.0"
eventlet = "^0.38.1"
django-celery-beat = "^2.7.0"


[tool.poetry.group.dependence_doc.dependencies]
drf-yasg = "^1.21.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
